جاري العمل على المرحلة الثانية من اللغة
جاري فصل ملف المحلل وإنشاء ملف مترجم
